pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "torch_uncertainty"
version = "0.5.0"
authors = [
    { name = "ENSTA U2IS", email = "olivier.laurent@ensta-paris.fr" },
    { name = "Adrien Lafage", email = "adrienlafage@outlook.com" },
    { name = "Olivier Laurent", email = "olivier.laurent@ensta-paris.fr" },
]
description = "Uncertainty quantification library in PyTorch"
readme = "README.md"
requires-python = ">=3.10"
keywords = [
    "bayesian-network",
    "ensembles",
    "neural-networks",
    "predictive-uncertainty",
    "reliable-ai",
    "trustworthy-machine-learning",
    "uncertainty",
    "uncertainty-quantification",
]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "timm",
    "lightning[pytorch-extra]>=2.0",
    "torchvision>=0.16",
    "einops",
    "seaborn",
]

[project.optional-dependencies]
experiments = ["tensorboard", "huggingface-hub>=0.31", "safetensors"]
image = ["kornia", "h5py", "opencv-python"]
tabular = ["pandas"]
dev = [
    "torch_uncertainty[experiments,image]",
    "ruff==0.11.10",
    "pytest-cov",
    "pre-commit",
    "pre-commit-hooks",
]
docs = [
    "sphinx==7.4.7",
    "pydata-sphinx-theme",
    "sphinx-copybutton",
    "sphinx-gallery",
    "sphinx-design",
    "sphinx-codeautolink",
    "sphinx_design",
    "sphinxcontrib-sass",
]
all = [
    "torch_uncertainty[dev,docs,tabular]",
    "scikit-learn",
    "laplace-torch",
    "scipy",
    "glest==0.0.1a1",
]

[project.urls]
homepage = "https://torch-uncertainty.github.io/"
documentation = "https://torch-uncertainty.github.io/quickstart.html"
repository = "https://github.com/ENSTA-U2IS-AI/torch-uncertainty.git"

[tool.flit.module]
name = "torch_uncertainty"

[tool.ruff]
line-length = 100
target-version = "py310"

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "__pycache__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "venv",
]

[tool.ruff.lint]
extend-select = [
    "A",
    "ARG",
    "B",
    "BLE",
    "C4",
    "D",
    "ERA",
    "F",
    "FURB",
    "G",
    "I",
    "INT",
    "ISC",
    "ICN",
    "N",
    "NPY",
    "PERF",
    "PIE",
    "PT",
    "PTH",
    "PYI",
    "Q",
    "RET",
    "RUF",
    "RSE",
    "S",
    "SIM",
    "T20",
    "TC",
    "TCH",
    "TID",
    "TRY",
    "UP",
    "YTT",
]
ignore = [
    "ARG002",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D107",
    "D205",
    "D206",
    "ISC001",
    "N818",
    "N812",
    "PT011",
    "RUF012",
    "S101",
    "TRY003",
]
pydocstyle.convention = "google"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 80

[tool.coverage.run]
branch = true
include = ["*/torch-uncertainty/*"]
omit = ["*/tests/*", "*/datasets/*"]

[tool.coverage.report]
exclude_lines = [
    "coverage: ignore",
    "raise NotImplementedError",
    "raise ImportError",
]
ignore_errors = true
